refactor(deep_causality_discovery): split cdl_effect into focused modules; extract CdlBuilder
The cdl_effect module lumped together two distinct concerns — the CdlEffect
monad and the CdlBuilder pipeline entry point — and inlined the full typeclass
stack in one 207-line file. Separate them.

- Extract CdlBuilder (struct, Effect3 impl, pure/build) into its own module
  types/cdl_builder/, leaving cdl_effect to own only the effect itself.
- Split the typeclass instances out of cdl_effect/mod.rs into dedicated files:
  functor.rs (Functor), applicative.rs (Pure + Applicative), monad.rs (Monad).
  mod.rs now keeps just the CdlEffect<T> type, its inherent methods, and the
  CdlEffectWitness HKT/HKT3 identity.
- Wire cdl_builder into types/mod.rs and re-export it from lib.rs so
  CdlBuilder resolves unchanged at the crate root; repoint the one explicit-path
  import in cdl_with_no_data.rs to crate::types::cdl_builder::CdlBuilder.
