+++ deep_causality_discovery/src/types/cdl_effect/mod.rs
@@ -3,14 +3,16 @@
  * Copyright (c) 2023 - 2026. The DeepCausality Authors and Contributors. All Rights Reserved.
  */
 
-use crate::{CDL, CdlConfig, CdlError, CdlWarningLog, NoData};
-use deep_causality_haft::{
-    Applicative, Effect3, Functor, HKT, HKT3, LogAppend, Monad, NoConstraint, Pure, Satisfies,
-};
+use crate::{CdlError, CdlWarningLog};
+use deep_causality_haft::{HKT, HKT3, NoConstraint};
 
 use std::marker::PhantomData;
 
-// --- 3.2. CdlEffect Type Definition ---
+mod applicative;
+mod functor;
+mod monad;
+
+// --- CdlEffect Type Definition ---
 
 #[derive(Debug, Clone)]
 pub struct CdlEffect<T> {
@@ -40,6 +42,7 @@ impl<T> CdlEffect<T> {
         F: FnMut(T) -> CdlEffect<B>,
     {
         // Delegate to the Monad Witness implementation
+        use deep_causality_haft::Monad;
         CdlEffectWitness::<CdlError, CdlWarningLog>::bind(self, f)
     }
 }
@@ -60,7 +63,7 @@ impl<T: std::fmt::Display> CdlEffect<T> {
     }
 }
 
-// --- 3.1. HKT Witness Definitions ---
+// --- HKT Witness Definition ---
 
 /// A witness type that "fixes" the Error (E) and Warning Log (WLog) types
 /// for `CdlEffect`, implementing the `HKT` and `HKT3` traits.
@@ -69,6 +72,9 @@ impl<T: std::fmt::Display> CdlEffect<T> {
 ///
 /// It uses `PhantomData` to carry the fixed types without inhibiting standard traits
 /// and a phantom generic placeholder.
+///
+/// The `Functor`, `Pure`, `Applicative`, and `Monad` instances for this witness live in
+/// the sibling `functor`, `applicative`, and `monad` modules.
 pub struct CdlEffectWitness<E, WLog>(PhantomData<(E, WLog)>);
 
 // Implement HKT
@@ -85,122 +91,4 @@ where
 {
     // The associated type that this witness "witnesses"
     type Type<T> = CdlEffect<T>;
-}
-
-// The Builder struct connecting the Effect system to the Witness
-pub struct CdlBuilder;
-
-// Implement Effect3: Fixing the Error and Warning types for the system.
-impl Effect3 for CdlBuilder {
-    type Fixed1 = CdlError;
-    type Fixed2 = CdlWarningLog;
-    type HktWitness = CdlEffectWitness<Self::Fixed1, Self::Fixed2>;
-}
-
-// --- Monad Implementation ---
-
-// 1. Functor: fmap
-impl Functor<CdlEffectWitness<CdlError, CdlWarningLog>>
-    for CdlEffectWitness<CdlError, CdlWarningLog>
-{
-    fn fmap<A, B, Func>(m_a: CdlEffect<A>, f: Func) -> CdlEffect<B>
-    where
-        Func: FnMut(A) -> B,
-    {
-        // fmap expects FnMut
-        let f = f;
-        CdlEffect {
-            inner: m_a.inner.map(f),
-            warnings: m_a.warnings, // Warnings are preserved
-        }
-    }
-}
-
-// 2a. Pure Implementation
-impl Pure<CdlEffectWitness<CdlError, CdlWarningLog>> for CdlEffectWitness<CdlError, CdlWarningLog> {
-    fn pure<T>(value: T) -> CdlEffect<T>
-    where
-        T: Satisfies<NoConstraint>,
-    {
-        CdlEffect {
-            inner: Ok(value),
-            warnings: CdlWarningLog::default(),
-        }
-    }
-}
-
-// 2. Applicative: apply
-impl Applicative<CdlEffectWitness<CdlError, CdlWarningLog>>
-    for CdlEffectWitness<CdlError, CdlWarningLog>
-{
-    fn apply<A, B, Func>(
-        f_ab: CdlEffect<Func>, // The container holding the function
-        mut m_a: CdlEffect<A>, // The container holding the value
-    ) -> CdlEffect<B>
-    where
-        Func: FnMut(A) -> B,
-        A: Clone,
-        B: Satisfies<NoConstraint>,
-        Func: Satisfies<NoConstraint>,
-    {
-        let mut combined_warnings = f_ab.warnings;
-        // Append warnings from m_a
-        combined_warnings.append(&mut m_a.warnings);
-
-        let new_inner = match (f_ab.inner, m_a.inner) {
-            (Ok(mut func), Ok(val)) => Ok(func(val)),
-            (Err(e), _) => Err(e),
-            (_, Err(e)) => Err(e),
-        };
-
-        CdlEffect {
-            inner: new_inner,
-            warnings: combined_warnings,
-        }
-    }
-}
-
-// 3. Monad: bind
-impl Monad<CdlEffectWitness<CdlError, CdlWarningLog>>
-    for CdlEffectWitness<CdlError, CdlWarningLog>
-{
-    fn bind<A, B, Func>(m_a: CdlEffect<A>, f: Func) -> CdlEffect<B>
-    where
-        A: Satisfies<NoConstraint>,
-        B: Satisfies<NoConstraint>,
-        Func: FnMut(A) -> CdlEffect<B>,
-    {
-        let mut f = f;
-        match m_a.inner {
-            Err(e) => CdlEffect {
-                inner: Err(e),
-                warnings: m_a.warnings,
-            },
-            Ok(val) => {
-                let mut m_b = f(val);
-                let mut combined_warnings = m_a.warnings;
-                // Append warnings from the result of the bound function
-                combined_warnings.append(&mut m_b.warnings);
-
-                CdlEffect {
-                    inner: m_b.inner,
-                    warnings: combined_warnings,
-                }
-            }
-        }
-    }
-}
-
-impl CdlBuilder {
-    /// Helper to lift a value into the CdlEffect context (Pure)
-    pub fn pure<T>(value: T) -> CdlEffect<T> {
-        CdlEffectWitness::<CdlError, CdlWarningLog>::pure(value)
-    }
-
-    pub fn build() -> CdlEffect<CDL<NoData>> {
-        Self::pure(CDL {
-            state: NoData,
-            config: CdlConfig::default(),
-        })
-    }
 }
